backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
python-decouple==3.8
cryptography==41.0.7
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0

# Content Creation Dependencies
beautifulsoup4==4.12.2
lxml==4.9.3
openai==1.3.7
stability-sdk==0.8.1
ffmpeg-python==0.2.0